First two months of our first game
Where we got our game during the first two months of learning to program
a platformer in Unity.  There are quite a few player controller bugs and
most of the code is in one class but it's playable.  If we were to
continue working on this game in the future the code will definitely
need to be cleaned up before proceeding development along with the other
bugs that are currently in this alpha version.  There is also a playable
version include as an .exe.